+++ src/Rules/Ip.php
@@ -21,12 +21,10 @@ class Ip implements RuleValidateInterface
      * @var array Rule properties
      */
     public static $config = [
-        'class' => 'Ip',
         'full_class' => __CLASS__,
         'alias' => ['ip'],
         'args_count' => 0,
-        'args_type' => [],
-        'has_validate' => true
+        'args_type' => []
     ];
 
     /**